+++ src/main/java/com/bcc/persistence/PlanDAO.java
@@ -3,87 +3,23 @@ package com.bcc.persistence;
 import java.util.List;
 
 import com.bcc.domain.BoardVO;
-import com.bcc.domain.GrpAcceptVO;
-import com.bcc.domain.MemberVO;
-import com.bcc.domain.PlanMemberVO;
 import com.bcc.domain.PlanVO;
 
 public interface PlanDAO {
-
-	// 아이디 검색 결과 회원 목록 가져오기
-	public List<MemberVO> getSearchMemList(String id);
-
-	// 그룹 번호 계산
-	public Integer getMaxGrpNum();
-
 	// 플랜 생성
 	public void createPlan(PlanVO vo);
 
-	// 회원 라이센스 가져오기
-	public Integer getLicense(String id);
-
-	// 회원 그룹 초대받은 목록
-	public List<GrpAcceptVO> getGrpAcceptList(String receiver);
-
-	// 회원이 소속된 그룹 정보 저장
-	public void insertGrpMember(PlanMemberVO member);
-
-	// 그룹 초대 수락 리스트에서 제거
-	public void deleteInvitation(GrpAcceptVO vo);
-
 	// 소속된 플랜 리스트 가져오기
-	public List<PlanVO> getPlanList(String id);
-
-	// 그룹 이름 가져오기
-	public String getGrpName(int num);
-
-	// 그룹에 소속된 멤버 리스트 가져오기
-	public List<MemberVO> getGrpMemberList(int grp_num);
-
-	// 플랜 멤버 삭제
-	public void delPlanMem(PlanMemberVO vo);
-
-	// 방장 정보 가져오기
-	public String getLeader(int num);
-
-	// 해당 그룹의 초대 멤버 리스트
-	public List<GrpAcceptVO> getInvitingList(int grp_num);
-
-	// 그룹에 멤버 초대
-	public void inviteMember(GrpAcceptVO vo);
-
-	// 회원 이름 가져오기
-	public String getMemberName(String id);
-
-	// 초대 취소
-	public void inviteCancle(GrpAcceptVO vo);
+	public List<PlanVO> getPlanList(String id);	
 	
-	// 그룹 멤버인지 확인
-	public int checkGrpMember(PlanMemberVO vo);
+	// 전체 카테고리별 관광지 정보 가져오기
+	public List<BoardVO> getTourList(int t_category);
 
-	// 다음 방장 아이디 가져오기
-	public String getNextLeader(int grp_num);
+	// 카테고리별 관광지 검색 결과
+	public List<BoardVO> getTourSearch(BoardVO vo);
 
-	// 방장 새로 설정
-	public void updateLeader(PlanVO plan);
-
-	// 전체 관광지 정보 가져오기
-	public List<BoardVO> getTourList();
-
-	// 전체 맛집 정보 가져오기
-	public List<BoardVO> getRestaurantList();
-
-	// 관광지 검색 결과
-	public List<BoardVO> getTourSearch(String keyword);
-
-	// 맛집 검색 결과
-	public List<BoardVO> getRestaurantSearch(String keyword);
-
-	// 관광지 하나 정보 가져오기
+	// 카테고리별 관광지 하나 정보 가져오기
 	public BoardVO getTourInfo(int num);
-	
-	// 맛집 하나 정보 가져오기
-	public BoardVO getRestaurantInfo(int num);
 	
 	// 여행 경로 수정
 	public void modifyTourPlan(PlanVO vo);
@@ -93,8 +29,5 @@ public interface PlanDAO {
 
 	// 플랜 삭제
 	public void delPlan(int num);
-	
-	// 그룹의 초대 리스트 모두 삭제
-	public void deleteInvitingList(int grp_num);
-	
+
 }
