+++ src/main/resources/mappers/planMapper.xml
@@ -6,5 +6,53 @@
 
 <mapper namespace="com.bcc.mapper.planMapper">
 
+	<insert id="createPlan">
+		insert into plan(num, grp_name, leader)
+		values(#{num}, #{grp_name}, #{leader})
+	</insert>
 
+	<select id="getPlanList" resultType="PlanVO">
+		select b.num, grp_name, leader
+		from plan_member a join plan b
+		on a.grp_num = b.num
+		where id = #{id}
+		order by grp_num desc
+	</select>
+	
+	<select id="getTourList" resultType="BoardVO">
+		select num, title, thumbnail, lng, lat
+		from tourist_spot
+		where t_category = #{t_category}
+		order by num asc
+	</select>
+
+	<select id="getTourSearch" resultType="BoardVO">
+		select num, title, thumbnail, lng, lat
+		from tourist_spot
+		where title like concat('%', #{title}, '%') and t_category = #{t_category}
+		order by num asc
+	</select>
+
+	<select id="getTourInfo" resultType="BoardVO">
+		select num, title, thumbnail, lng, lat, t_category
+		from tourist_spot
+		where num = #{num}
+	</select>
+	
+	<update id="modifyTourPlan">
+		update plan
+		set writer = #{writer}, tour_plan = #{tour_plan}, tour_plan_acc = #{tour_plan_acc}, tour_date_start = #{tour_date_start}, tour_date_end = #{tour_date_end}, lastUpdate = now()
+		where num = #{num}
+	</update>
+	
+	<select id="getPlanInfo" resultType="PlanVO">
+		select *
+		from plan
+		where num = #{num}
+	</select>
+	
+	<delete id="delPlan">
+		delete from plan
+		where num = #{num}
+	</delete>
 </mapper>
